requirements.txt
aniso8601==9.0.1
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-RESTful==0.3.10
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
PyJWT==2.9.0
pymongo==4.8.0
pytz==2024.1
six==1.16.0
Werkzeug==3.0.3
zipp==3.19.2
